+++ src/api/TaskDTO.h
@@ -5,9 +5,9 @@
 #ifndef TODOLIST_TASKDTO_H
 #define TODOLIST_TASKDTO_H
 
-#include "memory_model/Task.h"
+#include "memory_model/data/Task.h"
 #include "memory"
-#include "memory_model/TaskID.h"
+#include "memory_model/data/TaskID.h"
 
 
 /*
