+++ src/memory_model/structure/LinkManager.cpp
