+++ src/memory_model/structure/LinkManagerInterface.h
