+++ src/utils/task_io/ConsoleTaskIO.h
@@ -2,15 +2,15 @@
 // Created by denis on 13.07.20.
 //
 
-#ifndef TODOLIST_TASK_IO_H
-#define TODOLIST_TASK_IO_H
+#ifndef TODOLIST_CONSOLETASKIO_H
+#define TODOLIST_CONSOLETASKIO_H
 
 #include "api/TaskDTO.h"
 #include <unordered_map>
 #include <iomanip>
 
 
-namespace task_io {
+namespace console_task_io {
 
     const std::unordered_map<TaskPriority, std::string> priorty_repr = {
             {TaskPriority::FIRST, "first"},
@@ -24,7 +24,7 @@ namespace task_io {
 std::ostream& operator << (std::ostream& out, TaskDTO task) {
     out << "Task id_ = " << task.getId() << std::endl;
     out << "  name : " << task.getName() << std::endl;
-    out << "  priority : " << task_io::priorty_repr.at(task.getPriority()) << std::endl;
+    out << "  priority : " << console_task_io::priorty_repr.at(task.getPriority()) << std::endl;
     out << "  date : " << task.getDate() << std::endl;
     out << "  label : " << task.getLabel() << std::endl;
     out << "  complete : " << std::boolalpha << task.isCompleted() << std::endl;
@@ -35,4 +35,4 @@ std::ostream& operator << (std::ostream& out, TaskDTO task) {
 
 
 
-#endif //TODOLIST_TASK_IO_H
+#endif //TODOLIST_CONSOLETASKIO_H
