+++ tests/TestTaskNode.cpp
@@ -1,5 +1,5 @@
 #include <gtest/gtest.h>
-#include "memory_model/TaskNode.h"
+#include "memory_model/structure/TaskNode.h"
 
 using namespace boost::gregorian;
